add helper methods in lib to make it easy to use for third party plugins
